Moved getLeaves method to PhylogeneticTreeNode.
This method doesn't fit in the FilterBubbles class. Also, by moving it to another class,
this gives the opportunity to mock this method for testing purposes.

In this commit I also gave the private methods in Filterbubbles a TestId.
