+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/model/IPhylogeneticTreeNode.java
@@ -1,5 +1,7 @@
 package nl.tudelft.pl2016gr2.model;
 
+import java.util.ArrayList;
+
 /**
  * This is an interface which contains all of the methods which must be implemented by a container
  * class of a phylogenetic tree.
@@ -71,4 +73,10 @@ public interface IPhylogeneticTreeNode {
    * @return the index of the child ndoe.
    */
   int getChildIndex(IPhylogeneticTreeNode child);
+  
+  /**
+   * Gets all the leaves from the current node.
+   * @return : an ArrayList of strings with the labels of the leaves.
+   */
+  ArrayList<String> getLeaves();
 }
